Add patch to allow overriding CMAKE_CXX_STANDARD and set it
Conda-forge abseil is currently built with c++17 and requires any
downstream users to do the same for ABI compatibility. Libprotobuf is
one of those downstreams, and on conda-forge it has been built with
c++17, except that its propagated CMake target only sets the
project-minimum of c++14. Thus with gnsssdr requiring c++14, the part
using abseil does not get built with c++17 and it fails (currently only
on OSX, but still, a general problem). As a workaround to fixing
libprotobuf, we can just bump gnsssdr to use c++17 globally.